// src/app/Interfaces/v-unidades-ninos-dto.ts
export interface VUnidadesNinosDTO {
    iCveRegistroPersona?: number;
    cCURP?: string;
    cPrimerApellido?: string;
    cSegundoApellido?: string;
    cNombre?: string;
    iCveUnica: number;
    iCveCatEstatusNino?: number;
    cDescEstatusUnidad?: string;
    cDiagnosticoIngreso?: string;
    dFechaIngreso?: Date;
    dFechaBaja?: Date;
}